requirements.txt
google-api-core~=2.8.2
google-api-python-client~=2.51.0
google-auth~=2.8.0
google-auth-httplib2~=0.1.0
google-auth-oauthlib~=0.5.2
googleapis-common-protos~=1.56.2
html2text~=2020.1.16
oauthlib~=3.2.0
pylint~=2.14.3
requests~=2.28.0
requests-oauthlib~=1.3.1